% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_aae.R
\docType{data}
\name{aae}
\alias{aae}
\title{Cleaned AAE Data from NCT00089661}
\format{
\subsection{\code{aae}}{

A data frame with 802 rows and 2 columns:
\describe{
\item{SUBJID}{Subject Identifier for the Study}
\item{AETOXGR}{Standard Toxicity Grade}
}
}
}
\usage{
aae
}
\description{
AAE Data for BIS 620 Final Project
}
\keyword{datasets}
