% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_abmdxa.R
\docType{data}
\name{abmdxa}
\alias{abmdxa}
\title{Cleaned "abmdxa" Data from NCT00089661}
\format{
\subsection{\code{abmdxa}}{

A data frame with 1840 rows and 4 columns:
\describe{
\item{SUBJID}{Subject Identifier for the Study}
\item{BMLOC}{Location Used for the Measurement}
\item{DXABMCB}{DXA BMC at BL (g)}
\item{DXAAREAB}{DXA Area at BL (cm^2)}
}
}
}
\usage{
abmdxa
}
\description{
ABMDXA Data for BIS 620 Final Project
}
\keyword{datasets}
