% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{accuracy}
\alias{accuracy}
\title{function name: accuracy
Returns the accuracy of the glm model which is fitted using a subset of
the WDBC data.}
\usage{
accuracy(x = data, subset = "mean")
}
\arguments{
\item{x}{a dataframe. It is by default the WDBC data, which
has a diagnosis column and multiple columns with breast tissue
characteristics.
for example, radius_mean records the mean of breast tissue radius}

\item{subset}{a column name suffix input as character. It is by default
'mean', which subsets all columns recording the mean of breast tissue
characteristics such as radius_mean.}
}
\value{
The accuracy, which is defined as the proportion of
correctly-predicted diagnosis. The glm model used to calculate the accuracy
is fitted using the subset of the WDBC data containing
all columns with suffix 'subset'.
}
\description{
function name: accuracy
Returns the accuracy of the glm model which is fitted using a subset of
the WDBC data.
}
\examples{
data(data)
accuracy(data, "mean")
}
