% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adae.R
\docType{data}
\name{adae}
\alias{adae}
\title{Dataset adae}
\format{
import original dataframe from adae
}
\source{
{this is a test file uploading adae dataset to the package}
}
\usage{
adae
}
\description{
Dataset adae
}
\examples{
data(adae)
}
\keyword{datasets}
