% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adlb.R
\docType{data}
\name{adlb}
\alias{adlb}
\title{My Dataset adlb}
\format{
A data frame with 125123 rows and 7 variables:
}
\usage{
adlb
}
\description{
Import adlb dataset
}
\examples{
data(adae)
}
\keyword{datasets}
