% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adls.R
\docType{data}
\name{adls}
\alias{adls}
\title{My Dataset adls}
\format{
A data frame with 29118 rows and 11 variables:
}
\usage{
adls
}
\description{
Import adls dataset
}
\examples{
data(adls)
}
\keyword{datasets}
