% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adrsp.R
\docType{data}
\name{adrsp}
\alias{adrsp}
\title{My Dataset adrsp}
\format{
A data frame with 10856 rows and 9 variables:
}
\usage{
adrsp
}
\description{
Import adrsp dataset
}
\examples{
data(adrsp)
}
\keyword{datasets}
