% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_prediction.R
\name{ae_prediction}
\alias{ae_prediction}
\title{Fit a Multi-class Classification Model and Show the Evaluation Result}
\usage{
ae_prediction(input_df = df)
}
\arguments{
\item{input_df}{the input data frame that will be used in Multi-class
Classification model fitting; the default input is the
cleaned "df" dataset in our package}
}
\value{
a list that contains: 1. the call of the fitted random forest model
2. a confusion matrix made based on the
fitted model
3. a calculated result of feature importance in
our fitted random forest model
}
\description{
This function fits a multi-class classification model on the input
data frame using the random forest algorithm (applying 10-fold cross validation)
}
\author{
Houmin Xing & Yi Ren
}
