% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_avs.R
\docType{data}
\name{avs}
\alias{avs}
\title{Cleaned "AVS" Data from NCT00089661}
\format{
\subsection{\code{avs}}{

A data frame with 96 rows and 14 columns:
\describe{
\item{VISIT}{Visit name}
\item{SUBJID}{The unique identifier assigned to each patient}
\item{AGEY}{Age at enrollment in years}
\item{SEX}{Sex}
\item{RACE}{Race}
\item{VSTESTCD}{Vital signs test short name}
\item{VSSTRESN}{Numerical results/finding in standard units}
}
}
}
\usage{
avs
}
\description{
AVS Data for BIS 620 Final Project
}
\keyword{datasets}
