% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestmodel_bic.R
\name{bestmodel_bic}
\alias{bestmodel_bic}
\title{Obtain the names of the significant estimators based on BIC criterion}
\usage{
bestmodel_bic(x)
}
\arguments{
\item{x}{an object inherited from data.frame. It is
assumed to have its first column being the response variable along with
all other columns as potential estimators}
}
\value{
a character value with names of significant estimators
}
\description{
Obtain the names of the significant estimators based on BIC criterion
}
