% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestmodel_cp.R
\name{bestmodel_cp}
\alias{bestmodel_cp}
\title{Obtain the names of the significant estimators based on CP criterion}
\usage{
bestmodel_cp(x)
}
\arguments{
\item{x}{an object inherited from data.frame. It is
assumed to have its first column being the response variable along with
all other columns as potential estimators}
}
\value{
a character value with names of significant estimators
}
\description{
Obtain the names of the significant estimators based on CP criterion
}
