% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomark.R
\docType{data}
\name{biomark}
\alias{biomark}
\title{My Dataset biomark}
\format{
A data frame with 935 rows and 19 variables:
}
\usage{
biomark
}
\description{
Import biomark dataset
}
\examples{
data(biomark)
}
\keyword{datasets}
