% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boost_model.R
\name{boost_model}
\alias{boost_model}
\title{boost_model}
\usage{
boost_model(y, train, test = NULL)
}
\arguments{
\item{y}{character response variable}

\item{train}{training data that includes y}

\item{test}{test data that has the same column names as the training data, it
is an optional parameter with NULL default and triggers evaluation otherwise}
}
\value{
if test is NULL, it returns a xgb.Booster object, if a test object is
added then a list of 6 objects. The first is a xgb.Booster object, the second
is a confusion matrix, the third is an importance matrix, the fourth is a
PRROC object for the ROC curve, the fifth is a PRROC object for a precision-
recall curve, the sixth is an roc object for the roc results
}
\description{
Fits an XGBoost model to the input data and returns the model
object
}
\examples{
\dontrun{
data(diabetes)
xg_fit <- boost_model("Diabetes_binary", diabetes)
}
}
