% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_plot.R
\name{box_plot}
\alias{box_plot}
\title{box_plot}
\usage{
box_plot(df)
}
\arguments{
\item{df}{a dataframe with all the studies and their values of our
specified feature and affected_rate}
}
\value{
a box plot indicating the patterns of our selected feature
}
\description{
This function generates a box plot for our selected feature.
}
