% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_columns.R
\name{check_columns}
\alias{check_columns}
\title{Check Whether a Data Frame Has the Columns Needed for Conducting the
Multi-class Classification regarding Adverse Events}
\usage{
check_columns(input_df)
}
\arguments{
\item{input_df}{the input data frame that will be examined for
"column completeness"}
}
\value{
a TRUE/FALSE result that indicates whether the input data frame
has all the columns we need
}
\description{
This function examines the input data frame and determines whether it has all
the necessary columns required for fitting the Multi-class Classification model
}
\author{
Houmin Xing
}
