% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_test.R
\name{chi_square_matrix}
\alias{chi_square_matrix}
\title{Chi-Square Test Matrix}
\usage{
chi_square_matrix(normalization_matrix)
}
\arguments{
\item{normalization_matrix}{A matrix contains the normalized values from two specific columns in the data set.}
}
\description{
The function is to function in R is designed to compute a matrix of Chi-squared test p-values for pairs of columns in a given input matrix.
}
