% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_test.R
\name{clean_data}
\alias{clean_data}
\title{Clean the data}
\usage{
clean_data(data, x, y)
}
\arguments{
\item{data}{The dataset in which the function operates.}

\item{x}{The name of the column in data containing terms or items to be analyzed.}

\item{y}{The name of the column containing categorical data against which the terms in x are to be counted.}
}
\description{
The function is to clean and summarize a given dataset.
}
