% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{clean_data_time_series}
\alias{clean_data_time_series}
\title{Time Series Data Cleanning}
\usage{
clean_data_time_series(
  data,
  col1,
  col1_expected,
  col2,
  col2_expected_1,
  col2_expected_2,
  col2_expected_3,
  col3,
  col4,
  col5
)
}
\arguments{
\item{data}{The dataset in which the function operates.}

\item{col1}{The data frame to be processed.}

\item{col1_expected}{The expected value in col1 for filtering rows.}

\item{col2}{The name of the second column to be used for additional filtering.}

\item{col2_expected_1}{Value in col2 to be excluded from the dataset.}

\item{col2_expected_2}{Value in col2 to be excluded from the dataset.}

\item{col2_expected_3}{Value in col2 to be excluded from the dataset.}

\item{col3}{The name of the column representing the start date of each entry.}

\item{col4}{The name of the column representing the end or completion date of each entry.}

\item{col5}{The name of the column representing the end or completion date of each entry.}
}
\description{
The function is to preprocess a time series dataset, focusing on data cleaning and aggregation tasks.
}
