% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_helperfunc.R
\name{cleaning_helperfunc}
\alias{cleaning_helperfunc}
\title{Data Cleaning Helper Function}
\usage{
cleaning_helperfunc(df)
}
\arguments{
\item{df}{A data frame that needs cleaning.}
}
\value{
A cleaned data frame with categorical variables converted to numeric,
specific variables of interest selected, and rows with missing values removed.
}
\description{
This function performs several cleaning operations on a data frame. It converts
categorical columns to numeric, focusing on specific variables of interest,
and removes rows with missing values. It also provides a summary of missing values
in the specified variables of interest.
}
\details{
The function first converts categorical columns (excluding specified ones)
to numeric by treating NA or blank values as 'unknown'. It then selects
variables of interest and removes rows with any missing values in these
variables. The function also prints out the number of missing values for
each variable of interest both before and after the cleaning process.
}
