% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{condi}
\alias{condi}
\title{Condition Data sampled from ctgov-trials}
\format{
\subsection{\code{condi}}{

A data frame with 1,000 rows and 4 columns:
\describe{
\item{id}{id of study}
\item{nct_id}{nct id}
\item{name}{name}
\item{downcase_name}{downcase name}
}
}
}
\usage{
condi
}
\description{
Toy condi data for BIS620
}
\keyword{datasets}
