% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{create_ts}
\alias{create_ts}
\title{Create Time Series Vectors}
\usage{
create_ts(mesh_term)
}
\arguments{
\item{mesh_term}{A data frame that is expected to contain at least two columns named start_date and completion_date. These columns should contain date information for each entry in the data frame.}
}
\description{
The function is to convert the aggregated monthly counts into a vector and to create a time series object (ts) from this vector.
}
