% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{Synthesized WDBC Data}
\description{
WDBC data
}
\details{
\code{diagnosis} is the category parameter which
states the diagnosis for the breast tissue;
}
\keyword{data}
