% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcess.R
\name{data_process}
\alias{data_process}
\title{process the data
This function does some data processing.}
\usage{
data_process(x)
}
\arguments{
\item{x}{an object inherited from list.}
}
\value{
data for linear models.
}
\description{
process the data
This function does some data processing.
}
