% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{data_processing}
\alias{data_processing}
\title{Clean and preprocess the panitumumab data}
\usage{
data_processing(lesion, lesion_name)
}
\arguments{
\item{lesion}{the joined panitumumab dataset}

\item{lesion_name}{one lesion site name from unique(lesion$LSSiTE). This will
result in new_ls column with the 2 categories, 'lesion_name' and 'Other'.}
}
\value{
the lesion data frame with 2 new columns added.
}
\description{
Clean and preprocess the panitumumab data
}
\examples{
data(lesion)
data_processing(lesion, "Liver")
}
