% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcess.R, R/test_DTHDY.R
\name{DataPlot}
\alias{DataPlot}
\title{process the data
This function does some data processing.}
\usage{
DataPlot(x)

DataPlot(x)
}
\arguments{
\item{x}{an object inherited from list.}
}
\value{
data for linear models.

data for linear models.
}
\description{
process the data
This function does some data processing.

process the data
This function does some data processing.
}
