% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcess.R
\name{DataProcess}
\alias{DataProcess}
\title{process the data
This function does some data processing.}
\usage{
DataProcess(x)
}
\arguments{
\item{x}{an object inherited from list.}
}
\value{
data for linear models.
}
\description{
process the data
This function does some data processing.
}
