% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_var.R
\name{descriptive_var}
\alias{descriptive_var}
\title{Summarize and Plot Distribution of a Selected Variable
in Clinical Trials Data
This function summarizes and plots distribution of a selected categorical
variable in each of the study arm(s).}
\usage{
descriptive_var(df, subject_id_col, var_col, treat_col)
}
\arguments{
\item{df}{an object inherited from data.frame. It is
assumed to have a column containing participants' subject ids, a column
containing the selected variable's information (should be categorical),
a column containing study arm information.}

\item{subject_id_col}{the name of the column that contains
participants' id information}

\item{var_col}{the name of the column that contains the variable of
interest}

\item{treat_col}{the name of the column that contains the study arm
information}
}
\value{
a list containing the following two objects:
\enumerate{
\item a table with distribution of variable of interest in each study arm
\item a bar chart presenting distribution of variable of interest with
study arms shown as side-by-side bars.
}
}
\description{
Summarize and Plot Distribution of a Selected Variable
in Clinical Trials Data
This function summarizes and plots distribution of a selected categorical
variable in each of the study arm(s).
}
\examples{
data(full_data)
descriptive_var(full_data, "SUBJID", "B_ECOG", "ATRT")
}
