% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_df.R
\docType{data}
\name{df}
\alias{df}
\title{Cleaned Data from NCT00089661}
\format{
\subsection{\code{df}}{

A data frame with 96 rows and 14 columns:
\describe{
\item{SUBJID}{The unique identifier assigned to each patient}
\item{max_aetoxgr}{The maximun adverse event grade}
\item{SEX}{Sex}
\item{age}{Age}
\item{BMI}{Body Mass Index}
\item{DIABP}{Diastolic Blood Pressure}
\item{PULSE}{Pulse Rate}
\item{SYSBP}{Systolic Blood Pressure}
\item{RESP}{Respiratory Rate}
\item{TEMP}{Temperature}
\item{race}{Race}
\item{FEMORAL NECK}{Bone mineral density at femoral neck}
\item{LUMBAR SPINE}{Bone mineral density at lumbar spine}
\item{TOTAL HIP}{Bone mineral density at total hip}
}
}
}
\usage{
df
}
\description{
Data for BIS 620 Final Project
}
\keyword{datasets}
