% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diabetes}
\alias{diabetes}
\title{Kaggle Diabetes Health Indicator Data}
\description{
A subset of the Kaggle diabetes data where:
}
\details{
\code{Diabetes_binary} a binary 0/1 for diabetes;
\code{HighBP} a binary 0/1 for having high blood ;
\code{HighChol} a binary 0/1 for having high cholesterol;
\code{CholCheck} a binary 0/1 if they had a cholesterol check in the last 5 years;
\code{BMI} a continuous variable for the patient's BMI;
\code{Smoker} a binary 0/1 for if they are a smoker;
\code{Stroke} a binary 0/1 for if they have had a stroke;
\code{HeartDiseaseorAttack} a binary 0/1 for if they have had a heart attack or
heart diesase;
\code{PhysActivity} a binary 0/1 for having worked out in the last 30 days;
\code{Fruits} a binary 0/1 for consuming 1 or more fruits a day ;
\code{Veggies} a binary 0/1 for comsuming 1 or more vegetables a day ;
\code{HvyAlcoholConsump} a binary 0/1 for 14 or more drinks a week for men or 7 or
more drinks a week for women;
\code{AnyHealthcare} a binary 0/1 for whether or not they have health coverage ;
\code{NoDocbcCost} a binary 0/1 for if they needed to visit the doctor within the
last 12 months but could not because of cost;
\code{GenHlth} a 1-5 scale for bad to good general health;
\code{MentHlth} a 1-30 continuous variable for number of days for which there has
been a mental health problem;
\code{PhysHlth} a 1-30 continuous variable for number of days for which there has
been a physical health problem;
\code{DiffWalk} a binary 0/1 for difficulty walking or using stairs;
\code{Sex} a binary 0/1 for sex where 0 is female;
\code{Age} a continuous variable for age;
\code{Education} a 1-6 scale for the level of highest education completed;
\code{Income} a 1-8 scale for the patient's income bracket;
}
\keyword{data}
