% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_csv_button.R
\name{downloadCSV}
\alias{downloadCSV}
\title{Print "Download a data frame as CSV"}
\usage{
downloadCSV(data, filename = "data.csv")
}
\arguments{
\item{data}{The data frame to be downloaded.}

\item{filename}{The name of the file to save.}
}
\description{
Print This function generates a download button
for saving a data frame as a CSV file.
}
