% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_rf_and_glm.R
\name{eval_rf_and_glm}
\alias{eval_rf_and_glm}
\title{eval_rf_and_glm}
\usage{
eval_rf_and_glm(train, test, model, type = "response")
}
\arguments{
\item{train}{training data that includes y}

\item{test}{test data that has the same column names as the training data, it
is an optional parameter with NULL default and triggers evaluation otherwise}

\item{model}{random forest or GLM to evaluate}

\item{type}{character type of data to predict ("prob" for random forest, and
response" for glm)}
}
\value{
data.frame counterfactual table
}
\description{
Run random forest or GLM model
}
\examples{
\dontrun{
data(diabetes)
rf_fit <- rf_model("Diabetes_binary", train, test, optimize = TRUE)
eval_rf_and_glm(rf_fit, type = "prob")

glm_fit <- glm_model("Diabetes_binary", train, test, optimize = "manual")
eval_rf_and_glm(glm_fit, type = "response")
}
}
