% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_xg.R
\name{eval_xg}
\alias{eval_xg}
\title{eval_glm}
\usage{
eval_xg(train, test, model, y_name = "Diabetes_binary")
}
\arguments{
\item{train}{training data that includes y}

\item{test}{test data that has the same column names as the training data, it
is an optional parameter with NULL default and triggers evaluation otherwise}

\item{model}{function XGBoost model to evaluate}

\item{y_name}{character name of response variable in data set}
}
\value{
data.frame counterfactual table
}
\description{
Run XGBoost model
}
\examples{
\dontrun{
data(diabetes)
rf_fit <- rf_model("Diabetes_binary", train, test, optimize = TRUE)
eval_rf_and_glm(rf_fit, type = "prob")

glm_fit <- glm_model("Diabetes_binary", train, test, optimize = "manual")
eval_rf_and_glm(glm_fit, type = "response")
}
}
