% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_preprocess.R
\name{event_preprocess}
\alias{event_preprocess}
\title{event_preprocess}
\usage{
event_preprocess(x)
}
\arguments{
\item{x}{a table that contains all the data of the total reported events}
}
\value{
a data frame with columns of nct_id, subjects_affected,
subjects_at_risk, affected_rate
}
\description{
This function conducts the data-preprocessing for our
target variable: reported events.
After converting the input table into a data frame,
it drops all rows that has at least one NA for either subjects_affected
or subjects_at_risk.
It then sum ups all the subjects_affected and subjects_at_risks
across all groups for each study.
It then adds a new column named "affected_rate" by
subjects_affected/subjects_at_risk, which is the target variable
Finally, it filters out all the abnormal values of affected_rates and
returns a clean data frame as desired.
}
