% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_categorical_name.R
\name{extract_categorical_name}
\alias{extract_categorical_name}
\title{extract_categorical_name}
\usage{
extract_categorical_name(string)
}
\arguments{
\item{string}{character name of model variable}
}
\value{
formatted categorical variable name
}
\description{
Formats categorical variable names for use in formula() and
as.formula, specifically when R/glm_model() extracts significant variables.
object
}
\examples{
\dontrun{
extract_categorical_name("Education10")
extract_categorical_name("Education10TRUE")
}
}
