% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_category_of_X.R
\name{filterTrialsByCategory_of_X}
\alias{filterTrialsByCategory_of_X}
\title{Print "Filter Dataset by self_identified category"}
\usage{
filterTrialsByCategory_of_X(data, selected_category)
}
\arguments{
\item{data}{The data frame containing the data.}

\item{selected_category}{The category of X to filter by.}
}
\value{
A data frame with trials matching the specified category.
}
\description{
Print This function filters the dataset
based on self-identified category of X.
}
