% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_interval.R
\name{filterTrialsByDate}
\alias{filterTrialsByDate}
\title{Print "Filter Trials by Date Interval"}
\usage{
filterTrialsByDate(data, start_date, end_date)
}
\arguments{
\item{data}{The data frame containing the trial data with date-time fields.}

\item{start_date}{The start date-time for filtering in "YYYY-MM-DD HH:MM:SS" format.}

\item{end_date}{The end date-time for filtering in "YYYY-MM-DD HH:MM:SS" format.}
}
\value{
A data frame with trials active within the chosen date-time range.
}
\description{
Print Filters the clinical trial dataset
based on the provided start and end dates.
}
