% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_specified_value.R
\name{filterTrialsBySpecified_value}
\alias{filterTrialsBySpecified_value}
\title{Print "Filter Data by a specific value"}
\usage{
filterTrialsBySpecified_value(data, specified_value)
}
\arguments{
\item{data}{The data frame containing the dataset.}

\item{specified_value}{The specific value to filter by based on X.}
}
\value{
A data frame with trials fulfilling the condition.
}
\description{
Print This function filters the dataset
based on a specified value on X.
}
