% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit1.R
\name{fit1}
\alias{fit1}
\title{Create a Poisson Regression Model.
This function creates a Poisson regression model:
E_POV ~ E_DISABL + E_LIMENG + E_MINRTY+ E_NOHSDP + E_SNGPNT + E_UNEMP + E_PCI}
\usage{
fit1(x)
}
\arguments{
\item{x}{an object inherited from svi data.frame.}
}
\value{
the summary table of the regression model, the pseudo r-squared,
the correlation matrix for the variables.
}
\description{
Create a Poisson Regression Model.
This function creates a Poisson regression model:
E_POV ~ E_DISABL + E_LIMENG + E_MINRTY+ E_NOHSDP + E_SNGPNT + E_UNEMP + E_PCI
}
\examples{
data(svi)
fit1(svi)
}
