% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit4.R
\name{fit4}
\alias{fit4}
\title{Create a Poisson Regression Model.
This function creates a Poisson regression model:
E_CASE ~ E_DISABL + E_LIMENG + E_NOHSDP + E_SNGPNT + E_UNEMP + E_POV + E_PCI}
\usage{
fit4(x, y)
}
\arguments{
\item{x}{an object inherited from svi data.frame.}

\item{y}{an object inherited from covid data.frame.}
}
\value{
the summary table of the regression model, the pseudo r-squared,
the correlation matrix for the variables.
}
\description{
Create a Poisson Regression Model.
This function creates a Poisson regression model:
E_CASE ~ E_DISABL + E_LIMENG + E_NOHSDP + E_SNGPNT + E_UNEMP + E_POV + E_PCI
}
\examples{
data(svi)
data(covid)
fit4(svi, covid)
}
