% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_logistic_regression.R
\name{fit_logistic_regression}
\alias{fit_logistic_regression}
\title{Fit Logistic Regression Model}
\usage{
fit_logistic_regression(train_data, test_data)
}
\arguments{
\item{train_data}{Training dataset including both predictors and response variable.}

\item{test_data}{Test dataset including both predictors and response variable.}
}
\value{
The fitted logistic regression model object (\code{glm} object).
}
\description{
Fits a logistic regression model to the training data and evaluates its performance on the test data.
The function prints the model's accuracy and Area Under the Curve (AUC) for the Receiver Operating
Characteristic (ROC) curve. It also returns the fitted model object.
}
\seealso{
\link[stats]{glm}, \link[pROC]{roc}
}
