% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_random_forest.R
\name{fit_random_forest}
\alias{fit_random_forest}
\title{Fit Random Forest Model}
\usage{
fit_random_forest(train_data, test_data)
}
\arguments{
\item{train_data}{Training dataset including predictors and response variable.}

\item{test_data}{Test dataset including predictors and response variable.}
}
\value{
The fitted random forest model object (\code{randomForest} object).
}
\description{
Fits a random forest model to the provided training data and evaluates its performance on the test data.
It prints the model's accuracy and generates a ROC curve with the AUC value. The function also plots
variable importance and returns the fitted random forest model object.
}
\seealso{
\link[randomForest]{randomForest}, \link[pROC]{roc}
}
