% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{full_data}
\alias{full_data}
\title{Synthesized Clinical Trail Data}
\description{
A data set from NCT00460265:
\code{ATRT} is the Actual Treatment;
\code{DTHDY} is the Day of Death;
\code{SUBJID} is the Subject ID of the patients;
\code{AGE} is the Age in Years at Screening;
\code{SEX} is the Sex;
\code{DIAGTYPE} is the Primary Tumor Diagnosis;
\code{HPV} is the HPV Status;
\code{B_ECOG} is the Baseline ECOG Performance Status;
\code{DIAGSTAG} is the Disease Stage at Diagnosis
\code{RACE} is the Race of patients;
\code{AETERM} is the adverse events;
\code{RSRESP} is the Overall Response Status;
\code{RSCONFYN} is the confirmed response
}
\keyword{data}
