% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ae.R
\name{get_ae}
\alias{get_ae}
\title{Summarize and Plot Distribution of Adverse Events in Clinical Trials Data
This function summarizes and plots distribution of adverse events in each
of the study arm(s).}
\usage{
get_ae(df, ae_col, treat_col, top = 20)
}
\arguments{
\item{df}{an object inherited from data.frame. It is
assumed to have a column containing participants' subject ids, a column
containing the selected variable's information (should be categorical),
a column containing study arm information.}

\item{ae_col}{the name of the column that contains adverse events}

\item{treat_col}{the name of the column that contains the study arm
information}

\item{top}{a numeric value represents the number of most frequent adverse
events to be shown in the table and the plot, default \code{20}}
}
\value{
a list containing the following two objects:
\enumerate{
\item a table with distribution of adverse events in each study arm
\item a plot presenting distribution of adverse events, faceted by study
arms
}
}
\description{
Summarize and Plot Distribution of Adverse Events in Clinical Trials Data
This function summarizes and plots distribution of adverse events in each
of the study arm(s).
}
\examples{
data(full_data)
get_ae(full_data, "AETERM", "ATRT", 20)
}
