% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetConcurrentTrials.R
\name{get_concurrent_trials}
\alias{get_concurrent_trials}
\title{Get the Number of Concurrent Trials for Each Date in a Set of Studies}
\usage{
get_concurrent_trials(d)
}
\arguments{
\item{d}{The studies to get the number of concurrent trials for.}
}
\value{
A tibble with a \code{date} column and a \code{count} of the number of
concurrent trials at that date.
}
\description{
This function computes the number of concurrent clinical trials for each date within a given set of studies.
It is particularly useful for understanding the distribution and overlap of clinical trials over time.
}
