% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_counterfactual.R
\name{get_counterfactual}
\alias{get_counterfactual}
\title{get_counterfactual}
\usage{
get_counterfactual(model, data)
}
\arguments{
\item{model}{fitted model for predictions (must be one of random forest, glm,
or XGBoost)}

\item{data}{diabetes data frame}
}
\value{
counterfactual dataframe, where rows are the variables of interest,
and columns are the predicted diabetic proportion before/after changing the
corresponding variable (note that the \code{Before} column is equivalent to the
model's accuracy because the subsets are from the diabetic population only)
}
\description{
Computes counterfactual table for diabetes data. A
counterfactual table compares how changes to certain variables affects the
predicted proportion of diabetes among the diabetic population. Specifically,
we are interested in the variables: \code{BMI}, \code{HeartDiseaseorAttack}, \code{HighBP},
\code{HighChol}, and \code{PhysActivity} variables. We adjust BMI to a range of 15-17,
which falls within the CDC's 5-85 percentile for ages 13 and under. The
latter 4 variables are binary categorical variables, and we adjust them from
unhealthy to healthy values, e.g., \code{PhysActivity == 0} (no regular
physical activity) to \code{PhysActivity == 1}.
}
\examples{
\dontrun{
data(diabetes)
rf_fit <- rf_model("Diabetes_binary", diabetes, optimize = TRUE)
counterfactual <- get_counterfactual(rf_fit[[1]], diabetes, TRUE)
}
}
