% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_response.R
\name{get_response}
\alias{get_response}
\title{Summarize and Plot Distribution of Best Response in Clinical Trials Data
This function summarizes and plots distribution of best response throughout
the treatment of a participant in each of the study arm(s) and selected
groups(optional).}
\usage{
get_response(
  df,
  subject_id_col,
  treat_col,
  response_col,
  response_conf_col = NULL,
  group_col = NULL
)
}
\arguments{
\item{df}{an object inherited from data.frame. It is
assumed to have a column containing participants' subject ids, a column
containing the selected variable's information (should be categorical),
a column containing study arm information.}

\item{subject_id_col}{the name of the column that contains
participants' id information}

\item{treat_col}{the name of the column that contains the study arm
information}

\item{response_col}{the name of the column that contains the response
to treatments. The column should contain only character types with
the values limited to \code{"Progressive disease"}, \code{"Stable disease"},
\code{"Partial response"}, \code{"Complete response"} in a case-insensitive form.
Leading and Trailing spaces are accepted}

\item{response_conf_col}{Optional. The name of the column that contains
information on whether the response is confirmed. Only \code{Y} values will be
considered as confirmed.If not provided, all responses in \code{response_col}
will be included. Default to \code{NULL}.}

\item{group_col}{Optional. The name of the column that contains grouping
information for counting. Can only be categorical with character type. If
provided, responses levels will be counted within each unique group.
Default to \code{NULL}.}
}
\value{
a list containing the following two objects:
\enumerate{
\item a table with distribution of best response in each study arm
\item a plot presenting distribution of best response, faceted by study
arms (if no \code{group_col} provided), or a plot presenting distribution of
best response, faceted by study arms and groups (if \code{group_col} provided)
}
}
\description{
Summarize and Plot Distribution of Best Response in Clinical Trials Data
This function summarizes and plots distribution of best response throughout
the treatment of a participant in each of the study arm(s) and selected
groups(optional).
}
\examples{
data(full_data)
get_response(full_data, "SUBJID", "ATRT", "RSRESP", "RSCONFYN",
"DIAGTYPE") #both `response_conf_col` and `group_col` given
get_response(full_data, "SUBJID", "ATRT", "RSRESP", "RSCONFYN") #only
#`response_conf_col`
get_response(df = full_data, subject_id_col = "SUBJID",
treat_col = "ATRT", response_col = "RSRESP", group_col = "DIAGTYPE")
#only `group_col` provided
get_response(df = full_data, subject_id_col = "SUBJID",
treat_col = "ATRT", response_col = "RSRESP") #no `response_conf_col`
#or `group_col`.
}
