% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct-util_mid.R
\name{get_term_matrix}
\alias{get_term_matrix}
\title{Get matrix of terms}
\usage{
get_term_matrix(x)
}
\arguments{
\item{x}{the studies to get the number of keywords for.}
}
\value{
a matrix containing words existed in the keyword column of x
and their frequency.
}
\description{
The function is to get the keywords frequency of
selected studies.
}
