% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elderly_trials.R
\name{get_trials_for_older_adults}
\alias{get_trials_for_older_adults}
\title{Extract Trials Involving Elderly Participants}
\usage{
get_trials_for_older_adults(con)
}
\arguments{
\item{con}{A database connection object.}
}
\value{
A data frame of trials involving older adults.
}
\description{
Retrieves clinical trials from the \code{ctgov.duckdb} database that are explicitly marked
as involving older adults.
}
