% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{\%!in\%}
\alias{\%!in\%}
\title{Plot VITAL Data
This function plots Vital data of a disease by groups.}
\usage{
x \%!in\% y
}
\arguments{
\item{x}{a particular disease in vital data frame}

\item{v}{a data frame}
}
\value{
a histogram of person years of a particular
disease faceted by four groups.
}
\description{
Plot VITAL Data
This function plots Vital data of a disease by groups.
}
\examples{
data(vital)
}
