% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{histogram}
\alias{histogram}
\title{function name: histogram
Returns the histogram of a column from the WDBC data.}
\usage{
histogram(x = data, col = "radius_mean")
}
\arguments{
\item{x}{a dataframe. It is by default the WDBC data, which
has a diagnosis column and multiple columns
with breast tissue characteristics.
for example, radius_mean records the mean of breast tissue radius}

\item{col}{a column name input as character.
It is by default 'radius_mean', which records the mean of
breast tissue radius}
}
\value{
The histogram of the column from the WDBC data
with corresponding sizes
}
\description{
function name: histogram
Returns the histogram of a column from the WDBC data.
}
\examples{
data(data)
histogram(data, 'radius_mean')
}
