% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchClinicalTrialsApp.R
\name{launchClinicalTrialsApp}
\alias{launchClinicalTrialsApp}
\title{Launch Clinical Trials Shiny Application}
\usage{
launchClinicalTrialsApp()
}
\description{
This function encapsulates the UI and server logic for a Shiny application
that enables users to interactively explore clinical trials data.
}
