% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_shinyapp.R
\name{launchMyApp}
\alias{launchMyApp}
\title{Launch the Shiny App}
\usage{
launchMyApp()
}
\description{
This function starts the Shiny application contained in the bis620.2023 package.
}
\examples{
bis620.2023::launchMyApp()
}
