% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lesion}
\alias{lesion}
\title{The clinical trial study data set}
\description{
Key variables include:
}
\details{
\code{ATRT} Actual Treatment;
\code{AESEVCD} the Grade/Severity Code of the adverse effect, ranges from 1 to 4.
\code{LSSITE}Lesion Site;
\code{SUBJID} patient Subject ID;
}
\keyword{data}
