% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_test.R
\name{line_chart_comparison}
\alias{line_chart_comparison}
\title{Line Charts for Normalization Matrix}
\usage{
line_chart_comparison(
  normalization_matrix,
  count_matrix,
  column,
  var_name,
  value_name
)
}
\arguments{
\item{normalization_matrix}{A matrix containing normalized data for pairwise comparison.}

\item{count_matrix}{A matrix containing count data for Chi Test}

\item{column}{The name of the column in the matrix used for sorting before generating comparisons.}

\item{var_name}{The name of the variable in the long format that will hold the original column names of the matrix.}

\item{value_name}{The name of the value variable in the long format.}
}
\description{
The function is to perform pairwise comparisons between the columns of a normalization matrix
and generates line charts for those comparisons that are statistically significant.
}
\details{
and generates line charts for those comparisons that are statistically significant.
}
