% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression.R
\name{linear_regression}
\alias{linear_regression}
\title{process the data
This function builds linear models.}
\usage{
linear_regression(x)
}
\arguments{
\item{x}{an object inherited from list.}
}
\value{
data for linear models.
}
\description{
process the data
This function builds linear models.
}
