% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_transform.R
\name{log_transform}
\alias{log_transform}
\title{Log Transform The SVI data
This function log transform one column of SVI data, and replace all NA, and
Infinity with 0,}
\usage{
log_transform(column_data)
}
\arguments{
\item{column_data}{one column of data from the SVI dataframe.}
}
\value{
the same column that are log transformed, and all NA, and infinity
are replaced with 0.
}
\description{
Log Transform The SVI data
This function log transform one column of SVI data, and replace all NA, and
Infinity with 0,
}
\examples{
data(svi_tb)
log_transform(svi_tb$E_SPAN)
}
