% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{logit_analysis}
\alias{logit_analysis}
\title{Perform a logistic regression analysis on the panitumumab data}
\usage{
logit_analysis(x, lesion_site)
}
\arguments{
\item{x}{must be the \code{lesion} dataset}

\item{lesion_site}{this must be one lesion site name from
unique(lesion$LSSITE).
Data will be processed with data_processing(lesion_name = colname).}
}
\value{
the summary result for the logistic regression model
}
\description{
The response variable is level, and the predictors are ATRT and new_ls
}
