% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_plot.R
\name{mean_plot}
\alias{mean_plot}
\title{mean_plot}
\usage{
mean_plot(df)
}
\arguments{
\item{df}{a dataframe with all the studies and their values of our specified
feature and affected_rate}
}
\value{
a mean plot indicating the patterns of our selected feature
}
\description{
This function generates a mean plot for our selected feature.
}
