% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqplot.R
\name{model_plot}
\alias{model_plot}
\title{Plot a Poisson Regression Model.
This function plots a Poisson regression model:}
\usage{
model_plot(x, y, model)
}
\arguments{
\item{x}{an object inherited from svi data.frame.}

\item{y}{an object inherited from covid data.frame.}

\item{model}{which model should be plotted.}
}
\value{
diagnostic plots of the regression model.
}
\description{
Plot a Poisson Regression Model.
This function plots a Poisson regression model:
}
\examples{
data(svi)
data(covid)
model_plot(svi, covid, model = 1)
}
