% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pie.R
\name{pie}
\alias{pie}
\title{function 1: pie chart
Plot Pie chart for the Breast Cancer Wisconsin (Diagnostic) Data Set (WDBC)
This function plots WDBC data.}
\usage{
pie(x = data, colname = "diagnosis", eq = `!=`, rhs = 2)
}
\arguments{
\item{x}{a dataframe. It is by default the WDBC data, which
has a diagnosis column and multiple columns with breast
tissue characteristics.}

\item{colname}{a column name input as character.}

\item{eq}{a builtin condition such as less than (<) or more than (>)}

\item{rhs}{the right-hand-side of the condition; used to subset the data}
}
\value{
a pie plot showing the number of benign/malignant diagnosis
in the subsetted data
}
\description{
function 1: pie chart
Plot Pie chart for the Breast Cancer Wisconsin (Diagnostic) Data Set (WDBC)
This function plots WDBC data.
}
\examples{
data(data)
pie(data, 'diagnosis',`!=`,2)
}
