% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct-util_mid.R
\name{pie_chart_mesh}
\alias{pie_chart_mesh}
\title{Pie Chart}
\usage{
pie_chart_mesh(x)
}
\arguments{
\item{x}{the studies to get the top 10 mesh_terms for.}
}
\description{
The function is to get a pie chart of top 10 mesh_terms.
}
