% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactively_plot_spectral.R
\name{plot_accel_interactive}
\alias{plot_accel_interactive}
\title{Plot 3DF Accelerometry Data Interactively}
\usage{
plot_accel_interactive(x, x_var = "time")
}
\arguments{
\item{x}{A tibble with columns time, X, Y, Z.}

\item{x_var}{The name of the column to use for the x-axis, defaults to "time".}
}
\value{
A Plotly interactive plot object, faceted by channel, showing the acceleration or energy (in the Fourier domain) of the data.
}
\description{
Plot 3DF Accelerometry Data Interactively
}
