% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_age.R
\name{plot_age}
\alias{plot_age}
\title{Plot Age Distribution in Clinical Trials Data
This function plots distribution of participants' ages in
each of the study arm(s) separately.}
\usage{
plot_age(df, subject_id_col, age_col, treat_col, age_interval = 10)
}
\arguments{
\item{df}{an object inherited from data.frame. It is
assumed to have a column containing participants' subject ids, a column
containing age information (can be categorical or numerical), a column
containing study arm information.}

\item{subject_id_col}{the name of the column that contains
participants' id information}

\item{age_col}{the name of the column that contains the age information}

\item{treat_col}{the name of the column that contains the study arm
information}

\item{age_interval}{a numerical value that controls the age interval
represented by each bar in the histogram, default is \code{10}}
}
\value{
an age distribution histogram faceted by study arms.
}
\description{
Plot Age Distribution in Clinical Trials Data
This function plots distribution of participants' ages in
each of the study arm(s) separately.
}
\examples{
data(full_data)
plot_age(full_data, "SUBJID", "AGE", "ATRT", 10)
}
