% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bic.R
\name{plot_bic}
\alias{plot_bic}
\title{Plot the BIC value for different number of estimators}
\usage{
plot_bic(x)
}
\arguments{
\item{x}{an object inherited from data.frame. It is
assumed to have its first column being the response variable along with
all other columns as potential estimators}
}
\value{
a plot to visualize the optimal size of linear model based on BIC
}
\description{
Plot the BIC value for different number of estimators
}
