% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_condition_histogram.R
\name{plot_condition_histogram}
\alias{plot_condition_histogram}
\title{plot condition histogram}
\usage{
plot_condition_histogram(x, condition = condi)
}
\arguments{
\item{x}{the database table.}

\item{condition}{the condition table. (default: condi; the tibble of the
condition table)}
}
\value{
a histogram of the conditions of the  trials
}
\description{
create a histogram of the conditions that trials are examining
returned by a brief title keyword search
}
