% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct-util_mid.R
\name{plot_conditions_histogram}
\alias{plot_conditions_histogram}
\title{Plot a histogram of conditions}
\usage{
plot_conditions_histogram(x)
}
\arguments{
\item{x}{the database table.}
}
\value{
a histogram of conditions of selected studies
}
\description{
The function is to plot a histogram of conditions of selected
studies. The number of conditions is the smaller one of 50 and number of
conditions with percentage>0.0001.
}
