% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct-util_mid.R
\name{plot_countries}
\alias{plot_countries}
\title{Plot Countries}
\usage{
plot_countries(x)
}
\arguments{
\item{x}{the studies to get the world map for.}
}
\description{
The function is to get the number of studies in
different countries
and visualize in a world map.
}
