% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cp.R
\name{plot_cp}
\alias{plot_cp}
\title{Plot the Mallow's CP value for different number of estimators}
\usage{
plot_cp(x)
}
\arguments{
\item{x}{an object inherited from data.frame. It is
assumed to have its first column being the response variable along with
all other columns as potential estimators}
}
\value{
a plot to visualize the optimal size of linear model based on CP
}
\description{
Plot the Mallow's CP value for different number of estimators
}
