% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_elderly_trial_completion_by_phase.R
\name{plot_elderly_trial_completion_by_phase}
\alias{plot_elderly_trial_completion_by_phase}
\title{Plot Completion Status by Phase for Elderly Trials}
\usage{
plot_elderly_trial_completion_by_phase(studies, eligibilities)
}
\arguments{
\item{studies}{Data frame from the Studies table.}

\item{eligibilities}{Data frame from the Eligibilities table.}
}
\description{
This function plots the completion status of clinical trials involving elderly participants, grouped by trial phase.
}
