% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lesion.R
\name{plot_lesion}
\alias{plot_lesion}
\title{Plot histogram given colname name wrapped by ATRT
Plots are genereated based on the lesion data in this dataset.}
\usage{
plot_lesion(x, lesion_site)
}
\arguments{
\item{x}{must be the \code{lesion} dataset}

\item{lesion_site}{this must be one lesion site name from
unique(lesion$LSSiTE).
Data will be processed with data_processing(lesion_name = colname).}
}
\value{
a set of bar plots facet wrapped by ATRT and new_ls
(result from the data_processing)
}
\description{
Plot histogram given colname name wrapped by ATRT
Plots are genereated based on the lesion data in this dataset.
}
\examples{
data(lesion)
plot_lesion(lesion, "Liver")
}
