% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_num_arms.R
\name{plot_num_arms}
\alias{plot_num_arms}
\title{plot number of arms pirchart}
\usage{
plot_num_arms(x)
}
\arguments{
\item{x}{the database table.}
}
\value{
a piechart plot of the number of arms of the trials
}
\description{
create a piechart plot of the number of arms returned by a
brief title keyword search.
}
