% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct-util_mid.R
\name{plot_phase_histogram}
\alias{plot_phase_histogram}
\title{Plot a histogram of different phases}
\usage{
plot_phase_histogram(x)
}
\arguments{
\item{x}{the database table.}
}
\description{
The function is to plot a histogram of phases of selected table
with a fixed x-axis.
}
