% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct-util_mid.R
\name{plot_states_us}
\alias{plot_states_us}
\title{Plot states}
\usage{
plot_states_us(x)
}
\arguments{
\item{x}{the studies to get the US map for.}
}
\description{
The function is to get the number of studies in different states
and visualize in an US map.
}
