% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_study_type_histogram.R
\name{plot_study_type_histogram}
\alias{plot_study_type_histogram}
\title{plot study type histogram}
\usage{
plot_study_type_histogram(x)
}
\arguments{
\item{x}{the database table.}
}
\value{
a histogram of the study types of the trials
}
\description{
create a histogram of the study types returned by a brief title keyword search.
}
