% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_var.R
\name{predictive_var}
\alias{predictive_var}
\title{Do logistic regression test to see if a Selected Variable
in Clinical Trials Data is a moderator of the effect of treatment
on the death outcome}
\usage{
predictive_var(
  df,
  subject_id_col,
  death_col,
  treat_col,
  var_col = NULL,
  interaction = TRUE
)
}
\arguments{
\item{df}{an object inherited from data.frame. It is
assumed to have a column containing participants' subject ids, a column
containing the selected variable's information (should be categorical),
a column containing study arm information,
a column containing death flag information.}

\item{subject_id_col}{the name of the column that contains
participants' id information}

\item{death_col}{the name of the column that contains survival information.
Should be formatted as binary (factor or character)}

\item{treat_col}{the name of the column that contains the study arm
information}

\item{var_col}{Optional. The name of the column that contains the variable of
interest. Should be of type numeric or character. If not provided, this
function will analyze the effect of treatment on death only. If provided,
this function will analyze the effects of both
the selected variable and treatment and whether the selected variable
is predictive. Default to \code{NULL}}

\item{interaction}{Optional. Whether the interaction effect between treatment
and selected variable should be analyzed. Default to \code{TRUE}.}
}
\value{
a logistic model
}
\description{
This function do the logistic regression test for the interaction
effect of a selected categorical variable
and study arm(s) on the death outcome
}
\examples{
data(full_data)
predictive_var(full_data, "SUBJID","DEATH_FLAG", "ATRT", "DIAGSTAG", TRUE)
}
