% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prognostic_var.R
\name{prognostic_var}
\alias{prognostic_var}
\title{Do logistic regression test to see if a Selected Variable
in Clinical Trials Data is a prognostic in terms of the death outcome.}
\usage{
prognostic_var(df, subject_id_col, treat_col, control_val, death_col, var_col)
}
\arguments{
\item{df}{an object inherited from data.frame. It is
assumed to have a column containing participants' subject ids, a column
containing the selected variable's information (should be categorical),
a column containing study arm information,
a column containing death flag information.}

\item{subject_id_col}{the name of the column that contains
participants' id information}

\item{treat_col}{the name of the column that contains the study arm
information}

\item{control_val}{a string representing the name of the control treatment}

\item{death_col}{the name of the column that contains survival information.
Should be formatted as binary (factor or character)}

\item{var_col}{the name of the column that contains the variable of
interest.Can be of character or numeric type.}
}
\value{
a logistic model
}
\description{
This function conducts logistic regression test between the death outcome
and a selected variable.
}
\examples{
data(full_data)
prognostic_var(full_data, "SUBJID", "ATRT",
"Chemotherapy", "DEATH_FLAG", "HPV")
}
