% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.R
\name{randforest}
\alias{randforest}
\title{Create a Random Forest Model.
This function creates a random forest model:
EP_POV ~ EP_DISABL + EP_LIMENG + EP_MINRTY + EP_NOHSDP + EP_SNGPNT}
\usage{
randforest(x)
}
\arguments{
\item{x}{an object inherited from svi data.frame.}
}
\value{
the summary table of the random forest model, the plot of the error
rate, the ROC table, and the AUC score.
}
\description{
Create a Random Forest Model.
This function creates a random forest model:
EP_POV ~ EP_DISABL + EP_LIMENG + EP_MINRTY + EP_NOHSDP + EP_SNGPNT
}
\examples{
data(svi)
randforest(svi)
}
