% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{regression_model}
\alias{regression_model}
\title{Get the specific details of how are the health condition, intervention type,
gender and age factors related to adverse events.}
\usage{
regression_model(x, y, z, type)
}
\arguments{
\item{x}{an object inherited from a table with columns \code{subjects_affected},
\code{subjects_at_risk}, \code{nct_id}}

\item{y}{an object inherited from a table with columns \code{nct_id}, \code{gender},
\verb{healthy volunteers}, \code{adult}, \code{child}, \code{older_adult}}

\item{z}{an object inherited from a table with columns \code{nct_id},
\code{intervention_type}}

\item{type}{the type of regression model (containing two types: \code{without} model without interactions & \code{with} model with interactions)}
}
\value{
regression model summary of the relative significance of factors to adverse event.
}
\description{
The adverse rate is calculated by dividing the number of adverse event
affected people by the number of people at risk of adverse event.
An adverse event is considered serious if the adverse rate is more than 10\%.
Why choose 10\%?
https://www.ncbi.nlm.nih.gov/books/NBK558963/#:~:text=Adverse\%20events\%20(AE)\%20frequently\%20occur\%20in\%20any\%20medical\%20system\%2C,in\%20ten\%20patients\%20are\%20affected.
}
