% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_model.R
\name{rf_model}
\alias{rf_model}
\title{rf_model}
\usage{
rf_model(y, train, test = NULL, optimize = FALSE)
}
\arguments{
\item{y}{response variable}

\item{train}{training data}

\item{test}{test data}

\item{optimize}{how to optimize the model}
}
\description{
rf_model
}
\examples{
\dontrun{
rf_model("Diabetes_binary", train, test, "stepAIC")
}
}
