% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{start_shiny}
\alias{start_shiny}
\title{Function to start the toy shiny app for homework 3}
\usage{
start_shiny()
}
\value{
A shiny app object
}
\description{
Function to start the toy shiny app for homework 3
}
