% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct-util_mid.R
\docType{data}
\name{studies}
\alias{studies}
\title{My Dataset}
\format{
A data frame with 471252 rows and 13 columns:
}
\usage{
studies
}
\description{
A dataset containing information on clinical trials
}
\keyword{studies}
