% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summary_table.R
\name{summarizeData}
\alias{summarizeData}
\title{Print "Summarize Data"}
\usage{
summarizeData(data)
}
\arguments{
\item{data}{The data frame to summarize.}
}
\description{
Print This function calculates summary statistics for a data frame.
}
