% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaselineCharasteristics.R
\name{summary_baseline}
\alias{summary_baseline}
\title{Plot summary of baseline characteristics}
\usage{
summary_baseline(dl)
}
\arguments{
\item{dl}{an object inherited from list. It is
assumed to have a \code{adsl} list name}
}
\value{
a data.frame of summary of baseline characteristics
including number of male, number female, mean of age,
std of age, mean of weight, std of weight, mean of height,
std of height, number of different race.
}
\description{
This function summary the baseline characteristics
conditioned on
"FOLFOX alone" and "Panitumumab + FOLFOX"
}
\examples{
data(ukb_accel)
summary_response(ukb_accel)
}
