% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KRAS_summary.R
\name{summary_kras}
\alias{summary_kras}
\title{Plot number of KRAS Status}
\usage{
summary_kras(dl)
}
\arguments{
\item{dl}{an object inherited from list. It is
assumed to have a \code{biomark} and \code{adsl} list name}
}
\value{
a plot of number of KRAS Status.
}
\description{
This function Create a plot of number of total mutant,
wild-type, and unknown KRAS counts conditioned on
"FOLFOX alone" and "Panitumumab + FOLFOX"
}
\examples{
data(ukb_accel)
summary_response(ukb_accel)
}
