% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_response.R
\name{summary_response}
\alias{summary_response}
\title{Plot summary of response}
\usage{
summary_response(dl)
}
\arguments{
\item{dl}{an object inherited from list. It is
assumed to have a \code{adae}, \code{adrsp}, \code{adsl} list name}
}
\value{
a plot of number of different types of response.
}
\description{
This function summary the number of different types of
response conditioned on
"FOLFOX alone" and "Panitumumab + FOLFOX"
}
\examples{
data(ukb_accel)
summary_response(ukb_accel)
}
