% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_sex.R
\name{summary_sex}
\alias{summary_sex}
\title{Plot number of sex}
\usage{
summary_sex(dl)
}
\arguments{
\item{dl}{an object inherited from list. It is
assumed to have a \code{adsl} list name}
}
\value{
a plot of number of females and males.
}
\description{
This function Create a plot of number of total females and males
}
\examples{
data(ukb_accel)
summary_sex(ukb_accel)
}
