% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{summary_stats}
\alias{summary_stats}
\title{summary_stats}
\usage{
summary_stats(x)
}
\arguments{
\item{x}{a dataframe with all the studies and their values of our specified feature and affected_rate}
}
\value{
a tibble of the count, mean and standard deviation of different groups in our specified feature.
}
\description{
This function creates the summary of statistics of the affected_rate with respect to each feature.
}
