% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_toxicity}
\alias{summary_toxicity}
\title{Summary the toxicity condition
This function summary the grade of toxicity.}
\usage{
summary_toxicity(x)
}
\arguments{
\item{x}{an object inherited from list. It is
assumed to have a \code{adae} list name along with
a \code{AESEVCD} data.frame}
}
\value{
a summary form of grade of toxicity.
}
\description{
Summary the toxicity condition
This function summary the grade of toxicity.
}
\examples{
data(ukb_accel)
summary_toxicity(ukb_accel)
}
