% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvivalATRT.R
\name{survival_ATRT}
\alias{survival_ATRT}
\title{Plot survival analysis}
\usage{
survival_ATRT(dl, detail = FALSE)
}
\arguments{
\item{ls}{an object inherited from list. It is
assumed to have \code{adsl} and 'biomark' list name}
}
\value{
a visualization of survival analysis on the 'Plots' with
p-value, a gusummary on the 'Viewer'
and a table of survival analysis result on
"FOLFOX alone" and "Panitumumab + FOLFOX"
}
\description{
This function returns a data summary by treatment(ATRT) in the Viewer,
provides the survival analysis result of
"FOLFOX alone" and "Panitumumab + FOLFOX" and visualize it.
}
\examples{
data(ukb_accel)
summary_response(ukb_accel)
}
