% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survival_ATRT_KRAS.R
\name{survival_ATRT_KRAS}
\alias{survival_ATRT_KRAS}
\title{Ruturn the result of survival analysis}
\usage{
survival_ATRT_KRAS(dl, name = "Mutant", detailed = FALSE)
}
\arguments{
\item{ls}{an object inherited from list. It is
assumed to have \code{adsl} and 'biomark' list name}
}
\value{
a visualization of survival analysis on the 'Plots' with
p-value, two tibbles of survival analysis result on
"FOLFOX alone" and "Panitumumab + FOLFOX" conditioned on type
of KRAS you choose, and the summary of survival analysis
}
\description{
This function returns a result of survival analysis
to assess the effect of
treatment conditioned on different type of KRAS and visualize it.
}
\examples{
data(ukb_accel)
summary_response(ukb_accel)
}
