% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survival_KRAS.R
\name{Survival_KRAS}
\alias{Survival_KRAS}
\title{survival analysis considering KRAS}
\usage{
Survival_KRAS(x)
}
\arguments{
\item{x}{an object inherited from list.}
}
\value{
survival analysis plot
}
\description{
survival analysis considering KRAS
}
