% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct-util_mid.R
\name{table_with_link}
\alias{table_with_link}
\title{Table with Links}
\usage{
table_with_link(x)
}
\arguments{
\item{x}{the studies to get the table.}
}
\value{
table containing nct_id(with link),brief title, start date, and
completion date.
}
\description{
The function is to get a table of selected studies'
nct_id(with link),brief title, start date, and completion date.
}
