% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_DTHDY.R
\name{test_dthdy}
\alias{test_dthdy}
\title{process the data
This function builds a linear model
with a dependent variable death day "DTHTY"
And also test the Multicollinearity and Heteroskedasticity of the variables}
\usage{
test_dthdy(x)
}
\arguments{
\item{x}{an object inherited from list.}
}
\value{
data for linear models.
}
\description{
process the data
This function builds a linear model
with a dependent variable death day "DTHTY"
And also test the Multicollinearity and Heteroskedasticity of the variables
}
