% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{time_series_analysis}
\alias{time_series_analysis}
\title{Time Series Analysis}
\usage{
time_series_analysis(clean_data, col, col_value, date1, date2, date3)
}
\arguments{
\item{clean_data}{A data frame that contains the dataset to be analyzed.}

\item{col}{The name of the column in clean_data used for filtering the data based on col_value.}

\item{col_value}{The specific value in the column col used to filter the data.}

\item{date1}{The start date for filtering the data.}

\item{date2}{The end date for filtering the data.}

\item{date3}{The future date for filtering the data.}
}
\description{
The function is to design for comprehensive time series analysis and forecasting based on specific criteria within a given dataset. Also, this function is to process the dataset, to perform time series analysis, to generate various plots for analysis, to conduct forecasting, and to save these visualizations and results.
}
