% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_2vars.R
\name{vis_2vars}
\alias{vis_2vars}
\title{vis_2vars}
\usage{
vis_2vars(d, y, x)
}
\arguments{
\item{d}{diabetes data frame}

\item{y}{an arbitrary variable in diabetes data frame}

\item{x}{an arbitrary variable in diabetes data frame}
}
\description{
vis_2vars
}
\examples{
\dontrun{
vis_2vars(df, 'Diabetes_binary', 'BMI')
}
}
