% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_dist.R
\name{vis_dist}
\alias{vis_dist}
\title{vis_dist}
\usage{
vis_dist(d)
}
\arguments{
\item{d}{a data frame}
}
\value{
a historgram and qqPlot for all continuous variables in given data
}
\description{
Visualize the distributions of the non categorical variables in
the data set
}
\examples{
\dontrun{
data(diabetes)
vis_dist(diabetes)
}
}
