% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_num.R
\name{vis_num}
\alias{vis_num}
\title{vis_num}
\usage{
vis_num(d)
}
\arguments{
\item{d}{a data frame}
}
\value{
a grid.arrange of two ggplots: one of side by side boxplots of the
raw data for the numeric variables and one of side by side boxplots of the
scaled data for the numeric variables such that they all lie between 0 and 1
}
\description{
Visualize the non categorical variables in the dataset with
boxplots. Shows both scaled and non scaled variables
}
\examples{
\dontrun{
data(diabetes)
vis_num(diabetes)
}
}
