% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{vital_model_smoke}
\alias{vital_model_smoke}
\title{Plot VITAL Data
This function models and plots the relationship between bmi and disease
with the effect of smkoing}
\usage{
vital_model_smoke(v, x)
}
\arguments{
\item{v}{a data frame}

\item{x}{a particular disease in vital data frame}
}
\value{
a scatter plot with regression line of a particular disease
faceted by four groups.
}
\description{
Plot VITAL Data
This function models and plots the relationship between bmi and disease
with the effect of smkoing
}
\examples{
data(vital)
}
