% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boost_model.R
\name{boost_model}
\alias{boost_model}
\title{boost_model}
\usage{
boost_model(y, train, test = NULL)
}
\arguments{
\item{y}{response variable}

\item{train}{training data}

\item{test}{test data}
}
\description{
boost_model
}
\examples{
\dontrun{
xg.model <- boost_model(y, train, test)
plot(xg.model[[5]], main="Out-Of-sample PR curve")
xgb.plot.importance(xg.model[[3]][1:20,])
}
}
