% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boost_model.R
\name{boost_model}
\alias{boost_model}
\title{boost_model}
\usage{
boost_model(train, test, y, evaluate = FALSE)
}
\arguments{
\item{train}{training data}

\item{test}{test data}

\item{y}{response variable}

\item{evaluate}{option to evaluate the model}
}
\description{
boost_model
}
\examples{
\dontrun{
xg.model <- boost_model(train, test, y, evaluate = TRUE)
plot(xg.model[[5]], main="Out-Of-sample PR curve")
xgb.plot.importance(xg.model[[3]][1:20,])
}
}
