% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_model.R
\name{glm_model}
\alias{glm_model}
\title{glm_model}
\usage{
glm_model(train, test, y, optimize = NA, evaluate = FALSE)
}
\arguments{
\item{train}{training data}

\item{test}{testing data}

\item{y}{response variable}

\item{optimize}{how to optimize the glm}

\item{evaluate}{option to evaluate the model}
}
\description{
glm_model
}
\examples{
\dontrun{
glm_model(df, optimize = "manual")
glm_model(train, test, y, "stepAIC", evaluate = TRUE)
}
}
