% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchClinicalTrialsApp.R
\name{launchClinicalTrialsApp}
\alias{launchClinicalTrialsApp}
\title{Launch Clinical Trials Shiny Application}
\usage{
launchClinicalTrialsApp(con)
}
\arguments{
\item{con}{A database connection object created by DBI::dbConnect().
This connection is used for accessing the clinical trials data
stored in a database.}
}
\description{
This function encapsulates the UI and server logic for a Shiny application
that enables users to interactively explore clinical trials data.
}
