% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPhaseHistogram.R
\name{plot_phase_histogram}
\alias{plot_phase_histogram}
\title{Create a Histogram of the Phases with a Uniform X-axis}
\usage{
plot_phase_histogram(x)
}
\arguments{
\item{x}{A tibble containing the study phases to be plotted.}
}
\value{
A ggplot object representing the histogram with a fixed set of phase values on the x-axis.
}
\description{
This function generates a histogram of the study phases from the clinical trials database.
It ensures that the x-axis of the histogram always displays a consistent set of phase values,
regardless of the data returned by the query. This uniformity allows for easier comparison
across different queries.
}
