% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_phase_histogram.R
\name{plot_phase_histogram}
\alias{plot_phase_histogram}
\title{plot phase histogram}
\usage{
plot_phase_histogram(x)
}
\arguments{
\item{x}{the database table.}
}
\value{
a histogram of the phase of the query
}
\description{
create a histogram of the phases returned by a brief title keyword search
}
