% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_model.R
\name{rf_model}
\alias{rf_model}
\title{rf_model}
\usage{
rf_model(train, test, y, optimize = FALSE, evaluate = FALSE)
}
\arguments{
\item{train}{training data}

\item{test}{test data}

\item{y}{response variable}

\item{optimize}{how to optimize the model}

\item{evaluate}{option to evaluate the model}
}
\description{
rf_model
}
\examples{
\dontrun{
rf_model(train, test, y, optimize = TRUE, evaluate = TRUE)
}
}
