% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{studies}
\alias{studies}
\title{Studies Data Resampled from ctgove-trials}
\format{
\subsection{\code{studies}}{

A data frame with 1,000 rows and 70 columns:
\describe{
\item{nct_id}{nct id}
\item{study_type}{study type}
\item{brief_title}{brief title}
\item{phase}{phase}
\item{number_of_arms}{number of arms}
}
}
}
\usage{
studies
}
\description{
Toy studies data for BIS620
}
\keyword{datasets}
