% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survival_KRAS.R
\name{survival_kras}
\alias{survival_kras}
\title{survival analysis considering KRAS}
\usage{
survival_kras(x)
}
\arguments{
\item{x}{an object inherited from list.}
}
\value{
survival analysis plot
}
\description{
survival analysis considering KRAS
}
