% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_2vars.R
\name{vis_2vars}
\alias{vis_2vars}
\title{vis_2vars}
\usage{
vis_2vars(d, x, y)
}
\arguments{
\item{d}{a data frame}

\item{x}{an arbitrary character variable in diabetes data frame}

\item{y}{an arbitrary character variable in diabetes data frame}
}
\value{
a ggplot depending on the types of variables. Returns a violin
plot for one continuous and one factor variable, returns a bar plot for two
factor variables, and returns a scatterplot for two continuous variables
}
\description{
Visualize any two variables in the dataset
}
\examples{
\dontrun{
data(diabetes)
vis_2vars(diabetes, "BMI", "MentHlth")
}
}
